Completed seperate out template in sock namespace
